+++ projects/start-os/web/ui/src/app/services/api/mock-patch.ts
@@ -191,6 +191,7 @@ export const mockPatchData: DataModel = {
         eth0: {
           name: null,
           secure: null,
+          wanIpOverride: null,
           type: 'inbound-outbound',
           portMap: {
             pcp: { supported: null, at: null },
@@ -213,6 +214,7 @@ export const mockPatchData: DataModel = {
         wlan0: {
           name: null,
           secure: null,
+          wanIpOverride: null,
           type: 'inbound-outbound',
           portMap: {
             pcp: { supported: null, at: null },
@@ -238,6 +240,7 @@ export const mockPatchData: DataModel = {
         wireguard1: {
           name: 'StartTunnel',
           secure: null,
+          wanIpOverride: null,
           type: 'inbound-outbound',
           portMap: {
             pcp: { supported: null, at: null },
@@ -263,6 +266,7 @@ export const mockPatchData: DataModel = {
         wireguard2: {
           name: 'Mullvad VPN',
           secure: null,
+          wanIpOverride: null,
           type: 'outbound-only',
           portMap: {
             pcp: { supported: null, at: null },