+++ shared-libs/ts-modules/shared/src/i18n/dictionaries/de.ts
@@ -820,4 +820,5 @@ export default {
   928: 'Aufsteigend',
   929: 'Absteigend',
   930: 'Dienste in diesem Register werden von der Start9-Community gepflegt und befinden sich in der Betaphase. Fehler sind zu erwarten. Die Installation erfolgt auf eigenes Risiko.',
+  933: 'Manuell festgelegt',
 } satisfies i18n